requirements.txt
Flask
qrcode
Pillow
PyPDF2
cloudinary
psycopg2-binary
pikepdf
opencv-python-headless
numpy
img2pdf
reportlab
openpyxl
Werkzeug==3.0.0
